+++ sidebar/src/components/UI/molecules/bucket_create_button/style.js
@@ -1,0 +1,10 @@
+const style = {
+    margin: '15px',
+    minWidth: '120px',
+    minHeight: '30px',
+    backgroundColor: 'blue',
+    color: 'white',
+    fontWeight: 'bold',
+};
+
+export default style;
