+++ sidebar/src/components/templates/user_info_template/index.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
 import UserInfo from '../../UI/organisms/user_info';
 import FollowButton from '../../UI/molecules/follow_button';
 import UserBuckets from '../../UI/organisms/user_buckets';
